+++ source/keep_stream_by_language/changelog.md
@@ -1,4 +1,7 @@
 
+**<span style="color:#56adda">0.0.12</span>**
+- add option to unset subtitle dispositions (so subtitles off by default upon playback)
+
 **<span style="color:#56adda">0.0.11</span>**
 - add option to keep streams with no language tags or undefined values